requirements.txt
aiohttp
asyncio
flask
python-dotenv
py-cord
